move udt_whitelist to ckb_chain
