+++ src/ckb_chain/actor.rs
@@ -54,7 +54,6 @@ impl Actor for CkbChainActor {
         config: Self::Arguments,
     ) -> Result<Self::State, ActorProcessingErr> {
         let secret_key = config.read_secret_key()?;
-
         let secp = secp256k1::Secp256k1::new();
         let pub_key = secret_key.public_key(&secp);
         let pub_key_hash = ckb_hash::blake2b_256(pub_key.serialize());
